Remove needless StdCommand wrapper
